- corrected moving of enums only
